requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
Flask==1.1.2
Flask-SocketIO==4.3.0
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
mutagen==1.44.0
pycparser==2.20
pycryptodomex==3.9.7
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
requests==2.23.0
six==1.14.0
spotipy==2.12.0
tzlocal==2.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.4.2
